+++ core-ng/src/main/java/core/framework/impl/validate/v2/ObjectValidator.java
@@ -1,0 +1,8 @@
+package core.framework.impl.validate.v2;
+
+/**
+ * @author neo
+ */
+public interface ObjectValidator {
+    void validate(Object instance, ValidationErrors errors, boolean partial);
+}