+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Change log
 ### 4.16.1 (8/29/2017 - )
+* validate: replace validation impl with dynamic code generation
 
 ### 4.16.0 (8/23/2017 - 8/29/2017)
 * api: add validation for GET/DELETE body type